Load generic plugins from spyder_
