+++ spyder_breakpoints/locale/breakpoints.pot
