+++ spyder_breakpoints/locale/es/LC_MESSAGES/breakpoints.mo
