+++ spyder_breakpoints/locale/fr/LC_MESSAGES/breakpoints.mo
