+++ spyder_breakpoints/locale/pt_BR/LC_MESSAGES/breakpoints.mo
