+++ spyder_breakpoints/locale/ru/LC_MESSAGES/breakpoints.mo
