+++ spyder_profiler/images/profiler.png
