+++ spyder_profiler/locale/fr/LC_MESSAGES/profiler.mo
