+++ spyder_profiler/locale/pt_BR/LC_MESSAGES/profiler.mo
