+++ spyder_profiler/locale/ru/LC_MESSAGES/profiler.mo
