+++ spyder_pylint/images/pylint.png
