+++ spyder_pylint/locale/es/LC_MESSAGES/pylint.po
