+++ spyder_pylint/locale/fr/LC_MESSAGES/pylint.mo
