+++ spyder_pylint/locale/pt_BR/LC_MESSAGES/pylint.mo
