+++ spyder_pylint/locale/pylint.pot
