+++ spyder_pylint/locale/ru/LC_MESSAGES/pylint.mo
